requirement.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.3
django-ckeditor==5.9.0
django-cors-headers==3.0.2
django-crispy-forms==1.9.0
django-filter==2.1.0
django-formtools==2.1
django-guardian==2.0.0
django-import-export==2.0.2
django-js-asset==1.2.2
django-reversion==3.0.7
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
drf-extensions==0.5.0
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
oauthlib==3.0.2
odfpy==1.4.1
openpyxl==3.0.3
Pillow==6.0.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.10.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tablib==1.1.0
uritemplate==3.0.0
urllib3==1.25.3
xlrd==1.2.0
xlwt==1.3.0
